Split the state in multiple slices
