+++ libs/secrets/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.1",
   "private": true,
   "dependencies": {
-    "@nx/webpack": "19.4.1",
+    "@nx/webpack": "19.4.2",
     "dotenv-webpack": "^8.1.0"
   }
 }
