+++ libs/windy-sounding/package.json
@@ -16,7 +16,7 @@
     "@sveltejs/vite-plugin-svelte": "^3.1.1",
     "date-fns": "^3.6.0",
     "geojson": "^0.5.0",
-    "preact": "10.22.1",
+    "preact": "^10.22.1",
     "react-redux": "^9.1.2",
     "svelte": "^4.2.18"
   }
